Add files via upload
Write a function myAdd.m which can add two matrices with different dimensions, with the following usage:

output = myAdd(a, b)
where a and b are given 2D matrices, with perhaps different dimensions. Hence, it is not always possible to perform the common matrix addition. As an extension, myAdd can extend the given matrices to have the same dimensions (with minimum padded zeros), then add them to create the final output. For instance, when a = [1 2 3; 4 5 6] and b = [5; 2], myAdd(a, b) should return [6 2 3; 6 5 6]. Moreover, the function do not support scalar expansion, so myAdd([1, 1], 2）returns [3, 1] instead of [3, 3].