rotation and reflection
Write a MATLAB function myTransform.m to rotate or reflect a set of 2D points. The usage of the function is
output=myTransform(vec, theta, mode)
where
vec: a 2-by-n vector representing n points in 2D space
theta: angle (counter-clockwise, in radian) for rotation or reflection. For rotation, theta is the rotation angle. For reflection, it is the angle of the reflection axis.
mode: 'rotate' for rotation, 'reflect' for reflection
output: a 2-by-n vector representing the given n points after rotation or reflection